requirements.txt
Django==5.2.4
Pillow==11.3.0
python-decouple==3.8
django-ckeditor==6.7.3
django-taggit==6.1.0
django-js-asset==3.1.2
